# docs/swagger.yaml
basePath: /api/v1
definitions:
  request.LoginRequest:
    properties:
      code:
        description: 验证码登录时必填
        type: string
      loginType:
        description: 1:验证码登录 2:密码登录
        enum:
        - 1
        - 2
        type: integer
      password:
        description: 密码登录时必填
        maxLength: 20
        minLength: 6
        type: string
      username:
        description: 用户名或者手机号
        type: string
    required:
    - loginType
    - username
    type: object
  response.LoginResponse:
    properties:
      token:
        type: string
    type: object
  response.UserResponse:
    properties:
      avatar:
        type: string
      id:
        type: integer
      nickname:
        type: string
      work_space_list:
        description: 工作空间列表
        items:
          $ref: '#/definitions/response.WorkSpaceResponse'
        type: array
    type: object
  response.WorkSpaceAppResponse:
    properties:
      created_at:
        description: 创建时间
        type: string
      description:
        description: 应用描述
        type: string
      id:
        description: 应用ID
        type: integer
      name:
        description: 应用名称
        type: string
      status:
        description: 应用状态
        type: integer
    type: object
  response.WorkSpaceResponse:
    properties:
      app_list:
        description: 应用列表
        items:
          $ref: '#/definitions/response.WorkSpaceAppResponse'
        type: array
      created_at:
        description: 创建时间
        type: string
      description:
        description: 工作空间描述
        type: string
      id:
        description: 工作空间ID
        type: integer
      name:
        description: 工作空间名称
        type: string
      status:
        description: 工作空间状态
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for Noah API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Noah API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 支持手机验证码、密码登录
      parameters:
      - description: 登录请求
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/request.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.LoginResponse'
      summary: 用户登录
      tags:
      - Auth
  /user:
    get:
      consumes:
      - application/json
      description: 获取用户信息
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/response.UserResponse'
      summary: 获取用户信息
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"